Checked in 01/06/2017 Ticket 209,215,273,283

+++ Invessence/modules/web/src/main/java/com/invessence/web/data/custody/td/BenefiaciaryDetails.java
@@ -23,7 +23,7 @@ public class BenefiaciaryDetails implements Serializable
    private String beneCitizenshiId;
    private String typeOfBeneficiary;
    private String perStripes;
-   private Double sharePerc;
+   private Long sharePerc;
    private Date created;
    private String createdBy;
    private Date updated;
@@ -204,12 +204,12 @@ public class BenefiaciaryDetails implements Serializable
       this.perStripes = perStripes;
    }
 
-   public Double getSharePerc()
+   public Long getSharePerc()
    {
       return sharePerc;
    }
 
-   public void setSharePerc(Double sharePerc)
+   public void setSharePerc(Long sharePerc)
    {
       this.sharePerc = sharePerc;
    }
